refactor suite mapping
