+++ src/map_suite/mapSuiteResult.js
@@ -1,0 +1,49 @@
+import uuidV4 from 'uuid/v4'
+
+export const MapSuiteResult = (isRoot, data) => {
+    let suite = {
+        'title': '',
+        'suites': [],
+        'tests': [],
+        'pending': [],
+        'root': isRoot,
+        'fullFile': '',
+        'file': '',
+        'passes': [],
+        'failures': [],
+        'skipped': [],
+        'hasTests': false,
+        'hasSuites': false,
+        'totalTests': 0,
+        'totalPasses': 0,
+        'totalFailures': 0,
+        'totalPending': 0,
+        'totalSkipped': 0,
+        'hasPasses': false,
+        'hasFailures': false,
+        'hasPending': false,
+        'hasSkipped': false,
+        'duration': 0,
+        'rootEmpty': data.rootEmpty,
+        '_timeout': 0,
+        'uuid': uuidV4(),
+        'hasBeforeHooks': false,
+        'beforeHooks': [],
+        'hasAfterHooks': false,
+        'afterHooks': []
+    }
+
+    if (!isRoot) {
+        suite.title = data.title
+
+        if (data.sanitizedCapabilities) {
+            suite.title = `${suite.title} (${data.sanitizedCapabilities})`
+        }
+
+        if (data._duration) {
+            suite.duration = data._duration
+        }
+    }
+
+    return suite
+}
